Implement Mr. Stanton's comments
