+++ src/common/heartbeat2/package.xml
@@ -3,7 +3,7 @@
 <package format="3">
   <name>heartbeat2</name>
   <version>0.0.0</version>
-  <description>establish a heartbeat between drivestation and ROV</description>
+  <description>Establish a heartbeat between drivestation and ROV</description>
   <maintainer email="kadin_le@mines.edu">kadin</maintainer>
   <license>Apache-2.0</license>
 
@@ -13,11 +13,9 @@
   <test_depend>ament_lint_common</test_depend>
 
   <depend>rclcpp</depend>
-  <depend>builtin_interfaces</depend>
-  <depend>std_msgs</depend>
   <depend>csm_common_interfaces</depend>
   <depend>std_srvs</depend>
-  <depend>boost/circular_buffer</depend>
+  <depend>boost</depend>
   
 
   <export>
